Changed: move issue #306 regression test to integration suite
Relocate the regression test for 3D flip-repair cycling from the flips
module to an integration test. This allows cargo test --release to
accurately verify construction against release-mode constants, which is
not possible within unit tests. Additionally, update documentation for
DelaunayTriangulation::new to reflect that shuffled retries are now
enabled in all build profiles.

Refs: #306
